Create helper for string elements.
