+++ modules/json_form_widget/json_form_widget.services.yml
@@ -1,6 +1,8 @@
 services:
   json_form.builder:
     class: \Drupal\json_form_widget\JsonFormBuilder
-    arguments: ['@dkan.metastore.schema_retriever']
+    arguments: ['@dkan.metastore.schema_retriever','@json_form.string_helper']
   json_form.value_handler:
     class: \Drupal\json_form_widget\JsonFormValueHandler
+  json_form.string_helper:
+    class: \Drupal\json_form_widget\JsonFormStringHelper
